[libsocialcache] Convert content type of downloaded image if necessary. Contributes to JB#50799
When an image is downloaded, check its mime type based on the content
data. If the content mime type doesn't match the given file extension,
convert the image data to the type specified by the extension.

This ensures the image can be displayed correctly in applications, as
some may rely on the file extension to render the image rather than
the file content type.